refactor: split product_model into different files
