+++ src/components/product/Product.stories.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import Product from './Product'
-import { OPTIONS_MODES } from '../../models/product_model'
 import { Formik } from 'formik'
+import { OPTIONS_MODES } from '../../models/product_constants'
 
 const parameters = [
   {
